star rating & break big part down into smaller part
