+++ componentFunction/backToTop.php
@@ -1,0 +1,5 @@
+<div id="back-top" class="back-top">
+  <a href="#top">
+    <i class="fa fa-angle-up" aria-hidden="true"></i>
+  </a>
+</div>
