+++ componentFunction/navTabs.php
@@ -1,0 +1,22 @@
+<!--nav-tabs-->
+<ul class="nav nav-tabs gray-bg" role="tablist">
+    <!-- Sell -->
+    <li role="presentation" class="active col-md-4">
+    <a href="#sell" aria-controls="sell" role="tab" data-toggle="tab">
+        Sell
+    </a>
+    </li>
+    <!-- Rent -->
+    <li role="presentation" class="col-md-4">
+    <a href="#rent" aria-controls="rent" role="tab" data-toggle="tab">
+        Rent
+    </a>
+    </li>
+    <!-- Swap -->
+    <li role="presentation" class="col-md-4">
+    <a href="#swap" aria-controls="swap" role="tab" data-toggle="tab">
+        Swap
+    </a>
+    </li>
+</ul>
+<!--/nav-tabs-->
