+++ componentFunction/notification.php
@@ -1,0 +1,35 @@
+<style>
+    .errorWrap {
+        padding: 10px;
+        margin: 0 0 20px 0;
+        background: #fff;
+        border-left: 4px solid #dd3d36;
+        -webkit-box-shadow: 0 1px 1px 0 rgba(0,0,0,.1);
+        box-shadow: 0 1px 1px 0 rgba(0,0,0,.1);
+    }
+    .succWrap{
+        padding: 10px;
+        margin: 0 0 20px 0;
+        background: #fff;
+        border-left: 4px solid #5cb85c;
+        -webkit-box-shadow: 0 1px 1px 0 rgba(0,0,0,.1);
+        box-shadow: 0 1px 1px 0 rgba(0,0,0,.1);
+    }
+</style>
+
+<?php 
+  if($error){
+    ?>
+    <div class="errorWrap">
+      <?php echo htmlentities($error); ?>
+    </div>
+    <?php 
+  } 
+  if($msg){
+    ?>
+    <div class="succWrap">
+      <?php echo htmlentities($msg); ?>
+    </div>
+    <?php
+  }	
+?>
