+++ componentFunction/starRating.php
@@ -5,9 +5,9 @@
 	<TITLE>PHP Dynamic Star Rating using jQuery</TITLE>
 	<style>
 		body{width:610;}
-		.demo-table {width: 100%;border-spacing: initial;margin: 20px 0px;word-break: break-word;table-layout: auto;line-height:1.8em;color:#333;}
+		.demo-table {border-spacing: initial;margin: 20px 0px;word-break: break-word;table-layout: auto;line-height:1.8em;color:#333;}
 		.demo-table th {background: #999;padding: 5px;text-align: left;color:#FFF;}
-		.demo-table td {border-bottom: #f0f0f0 1px solid;background-color: #ffffff;padding: 5px;}
+		.demo-table td {border-bottom: #f0f0f0 1px solid;background-color: #ffffff;padding: 5px;border:none}
 		.demo-table td div.feed_title{text-decoration: none;color:#00d4ff;font-weight:bold;}
 		.demo-table ul{margin:0;padding:0;}
 		.demo-table li{cursor:pointer;list-style-type: none;display: inline-block;color: #F0F0F0;text-shadow: 0 0 1px #666666;font-size:20px;}
@@ -39,9 +39,25 @@
 				}
 			});
 			$.ajax({
-			url: "add_rating.php",
-			data:'id='+id+'&star='+$('#tutorial-'+id+' #rating').val(),
-			type: "POST"
+        url: "add_rating.php",
+        data:'id='+id+'&star='+$('#tutorial-'+id+' #rating').val(),
+        type: "POST"
+			});
+		}
+
+		function insertRating(obj,itemId,userId,id) {
+			$('.demo-table #tutorial-'+id+' li').each(function(index) {
+				$(this).addClass('selected');
+				$('#tutorial-'+id+' #rating').val((index+1));
+				if(index == $('.demo-table #tutorial-'+id+' li').index(obj)) {
+					return false;	
+				}
+			});
+			$.ajax({
+        url: "add_rating.php",
+        // data:'id='+id+'&star='+$('#tutorial-'+id+' #rating').val(),
+        data:'itemId='+itemId+'&userId='+userId+'&star='+$('#tutorial-'+id+' #rating').val(),
+        type: "POST"
 			});
 		}
 
@@ -61,14 +77,14 @@
 <!--rating-->
 	<table class="demo-table">
 		<tbody>
-			<tr>
+			<!-- <tr>
 				<th><strong>Tutorials</strong></th>
-			</tr>
+			</tr> -->
 
 <?php
 // echo "abc";
   $vhid = intval($_GET['vhid']);
-  $id = 3;
+  $id = $userId;
 /* echo $vhid;
 echo $id; */
   $sql = "SELECT * 
@@ -91,7 +107,7 @@ echo $id;  */
   $cnt = 1;
 /* echo $vhid;
 echo $id; */
-echo $query->rowCount();
+// echo $query->rowCount();
   /* if ($query->rowCount() === 0) {
     $sql = "INSERT INTO rating
       (itemId,userId)
@@ -107,23 +123,22 @@ echo $query->rowCount();
   } */
   if ($query->rowCount() > 0) {
       foreach ($results as $result) {
+        
+          /* echo "1";
           echo "abc";
           echo $vhid;
           echo $id;
           echo $result->star;
           echo $result->id;
-          echo "cba";
+          echo "cba"; */
 ?>
-  <!--  -->
-  <!--  -->
-  <!--  -->
   <tr>
     <td valign="top">
       <!-- <div class="feed_title"><?php echo $tutorial["title"]; ?></div> -->
 
       <div id="tutorial-<?php echo $result->id; ?>">
         <!-- <input type="hidden" name="rating" id="rating" value="<?php echo $result->star; ?>" /> -->
-        <input type="text" name="rating" id="rating" value="<?php echo $result->star; ?>" />
+        <input type="hidden" name="rating" id="rating" value="<?php echo $result->star; ?>" />
         <ul onMouseOut="resetRating(<?php echo $result->id; ?>);">
           <?php
             for($i=1;$i<=5;$i++) {
@@ -151,5 +166,45 @@ echo $query->rowCount();
   <!--  -->
   <!--  -->
 
-<?php	}} ?>
+<?php	}}else{ ?>
+<?php	
+  $tempId = 100;
+  $itemId = $vhid;
+  $userId = $id;
+
+  /* echo "2";
+  echo $sql;
+  echo " id=" . $id;
+  echo " vhid=" . $vhid;
+
+  echo " rowCount=" . $query->rowCount(); */
+?>
+  <tr>
+    <td valign="top">
+      <div id="tutorial-<?php echo $tempId; ?>">
+        <!-- <input type="hidden" name="rating" id="rating" value="<?php echo $result->star; ?>" /> -->
+        <input type="hidden" name="rating" id="rating" value="0" />
+        <ul onMouseOut="resetRating(<?php echo $tempId; ?>);">
+          <?php
+            for($i=1;$i<=5;$i++) {
+              $selected = "";
+              if(!empty($result->star) && $i<=$result->star) {
+                $selected = "selected";
+              }
+          ?>
+
+          <li class='<?php echo $selected; ?>' 
+            onmouseover="highlightStar(this,<?php echo $tempId; ?>);"
+            onmouseout="removeHighlight(<?php echo $tempId; ?>);" 
+            onClick="insertRating(this,<?= $itemId ?>,<?= $userId ?>,<?= $tempId; ?>);"
+            
+          >
+              &#9733;
+          </li>  
+          <?php }  ?>
+        <ul>
+      </div>
+    </td>
+  </tr>
+<?php } ?>
 <!--/rating-->
