+++ index.php
@@ -308,7 +308,7 @@
     </div>
   </div>
 </section>
-  <!-- /Resent Cat --> 
+  <!-- /Recent Cat --> 
 <!-- /Body -->
 
 <!--Footer -->
@@ -316,7 +316,7 @@
 <!-- /Footer--> 
 
 <!--Back to top-->
-<div id="back-top" class="back-top"> <a href="#top"><i class="fa fa-angle-up" aria-hidden="true"></i> </a> </div>
+<?php include 'componentFunction/backToTop.php';?>
 <!--/Back to top--> 
 
 <!--Login-Form -->
@@ -325,14 +325,13 @@
 
 <!--Register-Form -->
 <?php include('includes/registration.php');?>
-
 <!--/Register-Form --> 
 
 <!--Forgot-password-Form -->
 <?php include('includes/forgotpassword.php');?>
 <!--/Forgot-password-Form --> 
 
-<!-- Scripts --> 
+<!--Scripts--> 
 <script src="assets/js/jquery.min.js"></script>
 <script src="assets/js/bootstrap.min.js"></script> 
 <script src="assets/js/interface.js"></script> 
@@ -343,8 +342,7 @@
 <!--Slider-JS--> 
 <script src="assets/js/slick.min.js"></script> 
 <script src="assets/js/owl.carousel.min.js"></script>
+<!--/Scripts--> 
 
 </body>
-
-<!-- Mirrored from themes.webmasterdriver.net/carforyou/demo/index.html by HTTrack Website Copier/3.x [XR&CO'2014], Fri, 16 Jun 2017 07:22:11 GMT -->
 </html>
